+++ lib/ui/BrowsePanel.dart
@@ -178,22 +178,24 @@ class _BrowseContentWidgetState extends State<BrowseContentWidget> with Notifica
   Widget build(BuildContext context) {
     List<Widget> sectionsList = <Widget>[];
     if (_contentCodes != null) {
+      int focusOrderNumber = 0;
       for (String code in _contentCodes!) {
         List<String>? entryCodes = _BrowseSection.buildBrowseEntryCodes(sectionId: code);
         if ((entryCodes != null) && entryCodes.isNotEmpty) {
-          sectionsList.add(_BrowseSection(
+          focusOrderNumber++;
+          sectionsList.add(FocusTraversalOrder(order: NumericFocusOrder(focusOrderNumber.toDouble()), child: _BrowseSection(
             sectionId: code,
             entryCodes: entryCodes,
             expanded: _isExpanded(code),
-            onExpand: () => _toggleExpanded(code),)
+            onExpand: () => _toggleExpanded(code),))
           );
         }
       }
     }
 
-    return Padding(padding: EdgeInsets.all(16), child:
+    return FocusTraversalGroup(policy: OrderedTraversalPolicy(), child: Padding(padding: EdgeInsets.all(16), child:
       Column(children: sectionsList,),
-    ) ;
+    )) ;
   }
 
   void _updateContentCodes() {
@@ -286,8 +288,7 @@ class _BrowseSection extends StatelessWidget {
 
   Widget _buildHeading(BuildContext context) {
     return Padding(padding: EdgeInsets.only(bottom: (expanded ? 0 : 4)), child:
-      InkWell(onTap: () => _onTapHeading(context), child:
-        Container(
+        Semantics(container: true, explicitChildNodes: true, label: '$_title: $_description', child: Container(
           decoration: BoxDecoration(color: Styles().colors.white, border: Border.all(color: Styles().colors.surfaceAccent, width: 1),),
           padding: EdgeInsets.only(left: 16),
           child: Column(children: [
@@ -297,13 +298,8 @@ class _BrowseSection extends StatelessWidget {
                   Text(_title, style: Styles().textStyles.getTextStyle("widget.title.regular.fat"))
                 )
               ),
-              Opacity(opacity: _hasFavoriteContent ? 1 : 0, child:
-                Semantics(label: 'Favorite' /* TBD: Localization */, button: true, child:
-                  InkWell(onTap: () => _onTapSectionFavorite(context), child:
-                    FavoriteStarIcon(selected: _isSectionFavorite, style: FavoriteIconStyle.Button,)
-                  ),
-                ),
-              ),
+              if (_hasFavoriteContent)
+                IconButton(onPressed: () => _onTapSectionFavorite(context), icon: FavoriteStarIcon(selected: _isSectionFavorite, style: FavoriteIconStyle.Button,), tooltip: 'Favorite' /* TBD: Localization */, splashColor: Colors.transparent, highlightColor: Colors.transparent, hoverColor: Colors.transparent, focusColor: Colors.transparent,),
             ],),
             Row(crossAxisAlignment: CrossAxisAlignment.end, children: [
               Expanded(child:
@@ -311,22 +307,23 @@ class _BrowseSection extends StatelessWidget {
                   Text(_description, style: Styles().textStyles.getTextStyle("widget.info.regular.thin"))
                 )
               ),
-              Semantics(
-                label: expanded ? Localization().getStringEx('panel.browse.section.status.colapse.title', 'Colapse') : Localization().getStringEx('panel.browse.section.status.expand.title', 'Expand'),
-                hint: expanded ? Localization().getStringEx('panel.browse.section.status.colapse.hint', 'Tap to colapse section content') : Localization().getStringEx('panel.browse.section.status.expand.hint', 'Tap to expand section content'),
-                button: true, child:
-                  Container(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 16), child:
-                    SizedBox(width: 18, height: 18, child:
-                      Center(child:
-                        _headingIcon
-                      ),
-                    )
-                  ),
-              ),
+              IconButton(
+                splashColor: Colors.transparent, highlightColor: Colors.transparent, hoverColor: Colors.transparent, focusColor: Colors.transparent,
+                tooltip: expanded
+                    ? Localization().getStringEx('panel.browse.section.status.colapse.title', 'Colapse')
+                    : Localization().getStringEx('panel.browse.section.status.expand.title', 'Expand'),
+                onPressed: () => _onTapHeading(context),
+                icon: Container(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 16), child:
+                  SizedBox(width: 18, height: 18, child:
+                    Center(child:
+                      _headingIcon
+                    ),
+                  )
+                ),
+              )
             ],)
           ],)
-        ),
-      ),
+        )),
     );
   }
 
@@ -359,8 +356,8 @@ class _BrowseSection extends StatelessWidget {
           ));
         }
       }
-      return entriesList.isNotEmpty ? Padding(padding: EdgeInsets.only(left: 24), child:
-        Padding(padding: EdgeInsets.only(bottom: 4), child: Column(children: entriesList))
+      return entriesList.isNotEmpty ? FocusTraversalGroup(policy: OrderedTraversalPolicy(), child: Padding(padding: EdgeInsets.only(left: 24), child:
+        Padding(padding: EdgeInsets.only(bottom: 4), child: Column(children: entriesList)))
       ) : Container();
   }
 
@@ -508,22 +505,31 @@ class _BrowseEntry extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return InkWell(onTap: () => _onTap(context), child:
+    return Semantics(label: _title, container: true, explicitChildNodes: true, child:
         Container(
           decoration: BoxDecoration(color: Styles().colors.white, border: Border.all(color: Styles().colors.surfaceAccent, width: 1),),
           padding: EdgeInsets.zero,
           child:
             Row(children: [
-              Opacity(opacity: (favorite != null) ? 1 : 0, child:
-                HomeFavoriteButton(favorite: favorite, style: FavoriteIconStyle.Button, prompt: true,),
+              Visibility(
+                visible: (favorite != null),
+                maintainSize: true,
+                maintainState: false,
+                child: ExcludeFocus(
+                    excluding: (favorite == null),
+                    child: HomeFavoriteButton(
+                      favorite: favorite,
+                      style: FavoriteIconStyle.Button,
+                      prompt: true,
+                    )),
               ),
               Expanded(child:
                 Padding(padding: EdgeInsets.symmetric(vertical: 14), child:
                   Text(_title, style: Styles().textStyles.getTextStyle("widget.title.regular.fat"),)
                 ),
               ),
-              Padding(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 18),
-                  child: _iconWidget),
+              IconButton(tooltip: 'Expand', onPressed: () => _onTap(context), icon: Padding(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 18),
+                  child: _iconWidget), splashColor: Colors.transparent, highlightColor: Colors.transparent, hoverColor: Colors.transparent, focusColor: Colors.transparent,),
             ],),
         )
     );