+++ lib/ui/home/HomeEvent2Widget.dart
@@ -35,7 +35,6 @@ import 'package:illinois/ui/events2/Event2Widgets.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
 import 'package:illinois/ui/home/HomeWidgets.dart';
 import 'package:illinois/ui/settings/SettingsPrivacyPanel.dart';
-import 'package:illinois/ui/widgets/FocusableSemanticsWidget.dart';
 import 'package:illinois/ui/widgets/SemanticsWidgets.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
@@ -94,12 +93,12 @@ class _HomeEvents2WidgetState extends State<HomeEvents2Widget> {
   
   Widget get _contentTypeBar => Row(children:List<Widget>.from(
     FavoriteContentType.values.map((FavoriteContentType contentType) => Expanded(child:
-      FocusableSemanticsWidget(child: HomeFavTabBarBtn(contentType.eventsTitle.toUpperCase(),
+      HomeFavTabBarBtn(contentType.eventsTitle.toUpperCase(),
         position: contentType.position,
         selected: _contentType == contentType,
         onTap: () => _onContentType(contentType),
       )
-    )),
+    ),
   )));
 
   void _onContentType(FavoriteContentType contentType) {