+++ lib/ui/home/HomeGroupsWidget.dart
@@ -8,7 +8,6 @@ import 'package:illinois/ui/accessibility/AccessiblePageView.dart';
 import 'package:illinois/ui/groups/GroupHome2Panel.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
 import 'package:illinois/ui/home/HomeWidgets.dart';
-import 'package:illinois/ui/widgets/FocusableSemanticsWidget.dart';
 import 'package:illinois/ui/widgets/SemanticsWidgets.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/group.dart';
@@ -74,12 +73,12 @@ class _HomeGroupsWidgetState extends State<HomeGroupsWidget> {
 
   Widget get _contentTypeBar => Row(children:List<Widget>.from(
     FavoriteContentType.values.map((FavoriteContentType contentType) => Expanded(child:
-      FocusableSemanticsWidget(child: HomeFavTabBarBtn(contentType.groupsTitle.toUpperCase(),
+      HomeFavTabBarBtn(contentType.groupsTitle.toUpperCase(),
         position: contentType.position,
         selected: _contentType == contentType,
         onTap: () => _onContentType(contentType),
       )
-    )),
+    ),
   )));
 
   void _onContentType(FavoriteContentType contentType) {