+++ lib/ui/onboarding2/Onboarding2Widgets.dart
@@ -1,7 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:illinois/service/Analytics.dart';
-import 'package:illinois/ui/widgets/FocusableSemanticsWidget.dart';
+import 'package:illinois/ui/widgets/SemanticsWidgets.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/triangle_painter.dart';
@@ -92,7 +92,7 @@ class Onboarding2ToggleButton extends StatelessWidget{
 
   @override
   Widget build(BuildContext context) {
-    return FocusableSemanticsWidget(onSelect: () {
+    return WebFocusableSemanticsWidget(onSelect: () {
           onTap!();
           anaunceChange();
         },